+++ app/src/main/kotlin/jp/co/yumemi/android/codeCheck/adapter/SearchListAdapter.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import jp.co.yumemi.android.codeCheck.repository.GitHubRepositoryItem
+import jp.co.yumemi.android.codeCheck.model.GitHubRepositoryItem
 import jp.co.yumemi.android.code_check.databinding.LayoutItemBinding
 
 class SearchListAdapter(
