+++ app/src/main/res/navigation/nav_graph.xml
@@ -21,6 +21,6 @@
         tools:layout="@layout/fragment_detail">
         <argument
             android:name="item"
-            app:argType="jp.co.yumemi.android.codeCheck.repository.GitHubRepositoryItem" />
+            app:argType="jp.co.yumemi.android.codeCheck.model.GitHubRepositoryItem" />
     </fragment>
 </navigation>
